Initial server commit(v0.0.1) (#14)
* Initial client commit(v0.0.1)

This version is really unstable. So, do not use this version for
official program.

All of sources are based on the `ocf_client/` sources. Comparing to
`ocf_client` this is more generic version.

And you must add the device control in this client. AND THIS IS NOT
MEMBER VARIABLE IS BOOLEAN TYPE. THIS USES STRING TYPE. PRECISELY, JSON
TYPE.

* Codacy: Not up to standards. - 1

* Codacy: Not up to standards. - 2

* Initial server commit(v0.0.1)

This version is really unstable. I check this server correctly works...
maybe?

* Codacy: Not up to standards. - 1

* Device type fixed

All device type to `oid.d.ams`. And post the data added.